Working calendar, also moved components to CustomCalendar folder
