+++ api/src/main/resources/application.properties
@@ -1,7 +1,5 @@
 spring.application.name=appointment-book
 
-server.port=8081
-
 spring.config.import=optional:classpath:.env[.properties]
 
 spring.data.mongodb.uri=${MONGO_URI}
