+++ docker-compose.yaml
@@ -1,15 +1,15 @@
 services:
   api:
-    image:
-      scarletbobcat/nail-art:api
+    image: scarletbobcat/nail-art:api
+    platform: linux/amd64
     ports:
       - "8080:8080"
     environment:
       - SPRING_PROFILES_ACTIVE=dev
 
   client:
-    image:
-      scarletbobcat/nail-art:client
+    image: scarletbobcat/nail-art:client
+    platform: linux/amd64
     ports:
       - "5173:5173"
     environment:
